Add file for helper functions
